Add initial setup for e-commerce backend and database
Includes entity, repository, and model classes for core features such as Bundle, Cart, DiscountCode, and related entities. Adds Docker configuration for MySQL and Redis, along with initial database schema and seed data.

+++ src/main/java/com/ecom/pujakriti/entity/Address.java
@@ -16,7 +16,7 @@ public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "AddressID")
-    private Long addressId;
+    private Integer addressId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "UserID", nullable = false)
