+++ src/main/java/com/ecom/pujakriti/entity/OrderItem.java
@@ -15,7 +15,7 @@ public class OrderItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "OrderitemID")
-    private Long orderItemId;
+    private Integer orderItemId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "OrderID", nullable = false)
