+++ src/main/java/com/ecom/pujakriti/repository/BundleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BundleRepository extends JpaRepository<Bundle, Long> {
+public interface BundleRepository extends JpaRepository<Bundle, Integer> {
 }
