+++ src/main/java/com/ecom/pujakriti/repository/PujaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PujaRepository extends JpaRepository<Puja, Long> {
+public interface PujaRepository extends JpaRepository<Puja, Integer> {
 }
